Module hydroflow_lang::graph

Graph representation stages for Hydroflow graphs.

Modules

graph_algorithms

ops

Structs

DiMulGraph

A directed multigraph where an vertex’s inbound and outbound edges are indexed.

FlatGraphBuilder

GraphEdgeId

ID to identify an edge.

GraphNodeId

ID to identify a node (operator or handoff) in HydroflowGraph.

GraphSubgraphId

ID to identify a subgraph in HydroflowGraph.

HydroflowGraph

A graph representing a hydroflow dataflow graph (with or without subgraph partitioning, stratification, and handoff insertion). This is a “meta” graph used for generating Rust source code in macros from Hydroflow surface sytnax.

OpInstGenerics

OperatorInstance

Enums

Color

Node

PortIndexValue

Helper struct for PortIndex which keeps span information for elided ports.

Functions

build_hfcode

eliminate_extra_merges_tees

Removes missing merges and tees. Must be applied BEFORE subgraph partitioning.

get_operator_generics

node_color

Determine op color based on in and out degree. If linear (1 in 1 out), color is None.

partition_graph