Module hydroflow_lang::graph::graph_algorithms

General graph algorithm utility functions

Functions

scc_kosaraju

Finds the strongly connected components in the graph. A strongly connected component is a subset of nodes that are all reachable by each other.

topo_sort

Topologically sorts a set of nodes. Returns a list where the order of Ids will agree with the order of any path through the graph.

topo_sort_scc

Computers the topological sort of the nodes of a possibly cyclic graph by ordering strongly connected components together.