Struct hydroflow_lang::graph::ops::OperatorConstraints

pub struct OperatorConstraints {

    pub name: &'static str,
    pub categories: &'static [OperatorCategory],
    pub hard_range_inn: &'static dyn RangeTrait<usize>,
    pub soft_range_inn: &'static dyn RangeTrait<usize>,
    pub hard_range_out: &'static dyn RangeTrait<usize>,
    pub soft_range_out: &'static dyn RangeTrait<usize>,
    pub num_args: usize,
    pub persistence_args: &'static dyn RangeTrait<usize>,
    pub type_args: &'static dyn RangeTrait<usize>,
    pub is_external_input: bool,
    pub has_singleton_output: bool,
    pub ports_inn: Option<fn() -> PortListSpec>,
    pub ports_out: Option<fn() -> PortListSpec>,
    pub input_delaytype_fn: fn(_: &PortIndexValue) -> Option<DelayType>,
    pub write_fn: WriteFn,

}

An instance of this struct represents a single hydroflow operator.

Fields
name: &'static str

Operator’s name.

categories: &'static [OperatorCategory]

Operator categories, for docs.

hard_range_inn: &'static dyn RangeTrait<usize>

Input argument range required to not show an error.

soft_range_inn: &'static dyn RangeTrait<usize>

Input argument range required to not show a warning.

hard_range_out: &'static dyn RangeTrait<usize>

Output argument range required to not show an error.

soft_range_out: &'static dyn RangeTrait<usize>

Output argument range required to not show an warning.

num_args: usize

Number of arguments i.e. operator(a, b, c) has num_args = 3.

persistence_args: &'static dyn RangeTrait<usize>

How many persistence lifetime arguments can be provided.

type_args: &'static dyn RangeTrait<usize>

How many generic type arguments can be provided.

is_external_input: bool

If this operator receives external inputs and therefore must be in stratum 0.

has_singleton_output: bool

If this operator has a singleton reference output. For stateful operators. If true, WriteContextArgs::singleton_output_ident will be set to a meaningful value in the Self::write_fn invocation.

ports_inn: Option<fn() -> PortListSpec>

What named or numbered input ports to expect?

ports_out: Option<fn() -> PortListSpec>

What named or numbered output ports to expect?

input_delaytype_fn: fn(_: &PortIndexValue) -> Option<DelayType>

Determines if this input must be preceeded by a stratum barrier.

write_fn: WriteFn

The operator’s codegen. Returns code that is emited is several different locations. See OperatorWriteOutput.