Struct hydroflow_lang::graph::FlatGraph

pub struct FlatGraph { /* private fields */ }

A graph representing a hydroflow dataflow graph before subgraph partitioning, stratification, and handoff insertion. I.e. the graph is a simple “flat” without any subgraph heirarchy.

May optionally contain handoffs, but in this stage these are transparent and treated like an identity operator.

Use Self::into_partitioned_graph() to convert into a subgraph-partitioned & stratified graph.

Implementations

impl FlatGraph

pub fn new() -> Self

Create a new FlatGraph, same as Default::default().

pub fn insert_node( &mut self, node: Node, varname_opt: Option<Ident> ) -> GraphNodeId

Insert a node, assigning the given varname.

pub fn insert_operator_instance( &mut self, node_id: GraphNodeId, op_inst: OperatorInstance )

pub fn insert_edge( &mut self, src: GraphNodeId, src_port: PortIndexValue, dst: GraphNodeId, dst_port: PortIndexValue ) -> GraphEdgeId

Insert an edge between nodes thru the given ports.

pub fn nodes(&self) -> Iter<'_, GraphNodeId, Node>

Iterator over (GraphNodeId, &Node) pairs.

pub fn successors( &self, src: GraphNodeId ) -> impl '_ + Iterator<Item = (&PortIndexValue, GraphNodeId)> + DoubleEndedIterator + FusedIterator + Clone + Debug

Successors, iterator of (&PortIndexValue, GraphNodeId) of outgoing edges. PortIndexValue for the port coming out of src.

pub fn predecessors( &self, dst: GraphNodeId ) -> impl '_ + Iterator<Item = (&PortIndexValue, GraphNodeId)> + DoubleEndedIterator + FusedIterator + Clone + Debug

Predecessors, iterator of (&PortIndexValue, GraphNodeId) of incoming edges. PortIndexValue for the port going into dst.

pub fn surface_syntax_string(&self) -> String

Convert back into surface syntax.

pub fn degree_in(&self, dst: GraphNodeId) -> usize

Degree into a node.

pub fn degree_out(&self, src: GraphNodeId) -> usize

Degree out of a node.

pub fn write_surface_syntax(&self, write: &mut impl Write) -> Result

Convert back into surface syntax.

pub fn mermaid_string(&self) -> String

Convert into a mermaid graph.

pub fn write_mermaid(&self, write: &mut impl Write) -> Result

Convert into a mermaid graph.

pub fn into_partitioned_graph(self) -> Result<PartitionedGraph, Diagnostic>

Run subgraph partitioning and stratification and convert this graph into a PartitionedGraph.