Module hydroflow_lang::graph

Graph representation stages for Hydroflow graphs.

Modules

graph_algorithms

General graph algorithm utility functions

ops

Hydroflow’s operators

propagate_flow_props

Module for determining flow properties. See propagate_flow_props.

Structs

DiMulGraph

A directed multigraph where an vertex’s inbound and outbound edges are indexed.

FlatGraphBuilder

Wraper around HydroflowGraph to build a flat graph from AST code.

FlowProps

Stream and lattice properties. Used to determine correctness for scaling transformations.

GraphEdgeId

ID to identify an edge.

GraphNodeId

ID to identify a node (operator or handoff) in HydroflowGraph.

GraphSubgraphId

ID to identify a subgraph in HydroflowGraph.

HydroflowGraph

A graph representing a Hydroflow dataflow graph (with or without subgraph partitioning, stratification, and handoff insertion). This is a “meta” graph used for generating Rust source code in macros from Hydroflow surface sytnax.

OpInstGenerics

Operator generic arguments, split into specific categories.

OperatorInstance

Meta-data relating to operators which may be useful throughout the compilation process.

WriteConfig

Configuration for writing graphs.

Enums

Color

Push, Pull, Comp, or Hoff polarity.

GraphEdgeType

The type of the Hydroflow graph edge.

GraphNode

A node, corresponding to an operator or a handoff.

LatticeFlowType

Type of lattice flow.

PortIndexValue

Helper struct for PortIndex which keeps span information for elided ports.

WriteGraphType

Enum for choosing between mermaid and dot graph writing.

Functions

build_hfcode

The main function of this module. Compiles a HfCode AST into a HydroflowGraph and source code, or Diagnostic errors.

eliminate_extra_unions_tees

Removes missing unions and tees. Must be applied BEFORE subgraph partitioning, i.e. on a flat graph.

get_operator_generics

Gets the generic arguments for the operator. This helper method is here due to the special handling of persistence lifetimes ('static, 'tick, 'mutable) which must come before other generic parameters.

partition_graph

Main method for this module. Partions a flat HydroflowGraph into one with subgraphs.