Struct hydroflow_lang::graph::ops::OperatorConstraints

pub struct OperatorConstraints {
    pub name: &'static str,
    pub hard_range_inn: &'static dyn RangeTrait<usize>,
    pub soft_range_inn: &'static dyn RangeTrait<usize>,
    pub hard_range_out: &'static dyn RangeTrait<usize>,
    pub soft_range_out: &'static dyn RangeTrait<usize>,
    pub num_args: usize,
    pub is_external_input: bool,
    pub ports_inn: Option<&'static dyn Fn() -> PortListSpec>,
    pub ports_out: Option<&'static dyn Fn() -> PortListSpec>,
    pub input_delaytype_fn: &'static dyn Fn(&PortIndexValue) -> Option<DelayType>,
    pub write_fn: WriteFn,
}

Fields
name: &'static str

Operator’s name.

hard_range_inn: &'static dyn RangeTrait<usize>

Input argument range required to not show an error.

soft_range_inn: &'static dyn RangeTrait<usize>

Input argument range required to not show a warning.

hard_range_out: &'static dyn RangeTrait<usize>

Output argument range required to not show an error.

soft_range_out: &'static dyn RangeTrait<usize>

Output argument range required to not show an warning.

num_args: usize

Number of arguments i.e. operator(a, b, c) has num_args = 3.

is_external_input: bool

If this operator receives external inputs and therefore must be in stratum 0.

ports_inn: Option<&'static dyn Fn() -> PortListSpec>

What named or numbered input ports to expect?

ports_out: Option<&'static dyn Fn() -> PortListSpec>

What named or numbered output ports to expect?

input_delaytype_fn: &'static dyn Fn(&PortIndexValue) -> Option<DelayType>

Determines if this input must be preceeded by a stratum barrier.

write_fn: WriteFn

Emit code in multiple locations. See OperatorWriteOutput.